MeshcatVisualizer: Tweaks to support caching mesh geometry on the zmq server
Adds an argument to the constructor to avoid deleting the entire tree on every run.
Sets uuids deterministically based on the mesh file so that they can be recognized as identical on the server.
See https://github.com/rdeits/meshcat-python/pull/75

This makes a *dramatic* improvement in the workflow of using meshcat on colab (at least as we use it in drake), because we don't have to repeatedly download large meshfiles from repeated simulations.

+++ bindings/pydrake/systems/meshcat_visualizer.py
@@ -6,6 +6,7 @@ package, Meshcat:
 import argparse
 import math
 import os
+import uuid
 import warnings
 import webbrowser
 
@@ -214,6 +215,7 @@ class MeshcatVisualizer(LeafSystem):
                  frames_opacity=1.,
                  axis_length=0.15,
                  axis_radius=0.006,
+                 delete_prefix_on_load=True,
                  **kwargs):
         """
         Args:
@@ -242,6 +244,14 @@ class MeshcatVisualizer(LeafSystem):
                     {"1": {"A", "B"}}.
             frames_opacity, axis_length and axis_radius are the opacity, length
                 and radius of the coordinate axes to be drawn.
+            delete_prefix_on_load: Specifies whether we should delete the
+                visualizer path associated with prefix on our load
+                initialization event.  True ensures a clean slate for every
+                simulation.  False allows for the possibility of caching object
+                meshes on the zmqserver/clients, to avoid repeatedly
+                downloading meshes over the websockets link, but will cause
+                geometry from previous simulations to remain in the scene.  You
+                may call ``delete_prefix()`` manually to clear the scene.
 
         Additional kwargs will be passed to the MeshcatVisualizer constructor.
         Note:
@@ -253,6 +263,7 @@ class MeshcatVisualizer(LeafSystem):
         self.set_name('meshcat_visualizer')
         self.DeclarePeriodicPublish(draw_period, 0.0)
         self.draw_period = draw_period
+        self._delete_prefix_on_load = delete_prefix_on_load
 
         # Recording.
         self._is_recording = False
@@ -359,6 +370,16 @@ class MeshcatVisualizer(LeafSystem):
         frame_name = name[pos + len(delim):]
         return source_name, frame_name
 
+    def delete_prefix(self):
+        """
+        Manually delete the meshcat prefix path specified in the constructor.
+        All objects/transforms at or below this path will be removed.  This
+        effectively clears all previous visualizations.  You will need to call
+        ``load()`` directly, or trigger the initialization event that occurs at
+        the beginning of a simulation, to see any visualizations.
+        """
+        self.vis[self.prefix].delete()
+
     def load(self):
         """
         Loads ``meshcat`` visualization elements.
@@ -367,7 +388,8 @@ class MeshcatVisualizer(LeafSystem):
             The ``scene_graph`` used to construct this object must be part of a
             fully constructed diagram (e.g. via ``DiagramBuilder.Build()``).
         """
-        self.vis[self.prefix].delete()
+        if self._delete_prefix_on_load:
+            self.vis[self.prefix].delete()
 
         # Intercept load message via memq LCM.
         memq_lcm = DrakeLcm("memq://")
@@ -383,7 +405,7 @@ class MeshcatVisualizer(LeafSystem):
         # Translate elements to `meshcat`.
         for i in range(load_robot_msg.num_links):
             link = load_robot_msg.link[i]
-            [source_name, frame_name] = self._parse_name(link.name)
+            [_, frame_name] = self._parse_name(link.name)
 
             for j in range(link.num_geom):
                 geom = link.geom[j]
@@ -395,9 +417,27 @@ class MeshcatVisualizer(LeafSystem):
                 meshcat_geom, material, element_local_tf = _convert_mesh(geom)
                 if meshcat_geom is not None:
                     cur_vis = (
-                        self.vis[self.prefix][source_name][str(link.robot_num)]
+                        self.vis[self.prefix][str(link.robot_num)]
                         [frame_name][str(j)])
-                    cur_vis.set_object(meshcat_geom, material)
+                    # Make the uuid's deterministic for mesh geometry, to
+                    # support caching at the zmqserver.  This means that
+                    # multiple (identical) geometries may have the same UUID,
+                    # but testing suggests that meshcat + three.js are ok with
+                    # it.
+                    if isinstance(meshcat_geom, meshcat.geometry.MeshGeometry):
+                        meshcat_geom.uuid = str(uuid.uuid5(
+                            uuid.NAMESPACE_X500, meshcat_geom.contents))
+                        material.uuid = str(uuid.uuid5(
+                            uuid.NAMESPACE_X500, meshcat_geom.contents
+                            + "material"))
+                        mesh = meshcat.geometry.Mesh(meshcat_geom, material)
+                        mesh.uuid = str(uuid.uuid5(
+                            uuid.NAMESPACE_X500, meshcat_geom.contents
+                            + "mesh"))
+                        cur_vis.set_object(mesh)
+                    else:
+                        cur_vis.set_object(meshcat_geom, material)
+
                     cur_vis.set_transform(element_local_tf)
 
                 # Draw the frames in self.frames_to_draw.
@@ -408,8 +448,8 @@ class MeshcatVisualizer(LeafSystem):
                     link_name = frame_name
                 if (robot_name in self.frames_to_draw.keys()
                         and link_name in self.frames_to_draw[robot_name]):
-                    prefix = (self.prefix + '/' + source_name + '/'
-                              + str(link.robot_num) + '/' + frame_name)
+                    prefix = (self.prefix + '/' + str(link.robot_num) + '/'
+                              + frame_name)
                     AddTriad(
                         self.vis,
                         name="frame",
@@ -428,14 +468,14 @@ class MeshcatVisualizer(LeafSystem):
         for frame_i in range(pose_bundle.get_num_poses()):
             # SceneGraph currently sets the name in PoseBundle as
             #    "get_source_name::frame_name".
-            [source_name, frame_name] = self._parse_name(
+            [_, frame_name] = self._parse_name(
                 pose_bundle.get_name(frame_i))
             model_id = pose_bundle.get_model_instance_id(frame_i)
             # The MBP parsers only register the plant as a nameless source.
             # TODO(russt): Use a more textual naming convention here?
             pose_matrix = pose_bundle.get_transform(frame_i)
             cur_vis = (
-                self.vis[self.prefix][source_name][str(model_id)][frame_name])
+                self.vis[self.prefix][str(model_id)][frame_name])
             cur_vis.set_transform(pose_matrix.GetAsMatrix4())
             if self._is_recording:
                 with self._animation.at_frame(
